third_party/blink/web_tests/platform/linux/virtual/orb-v01/external/wpt/fetch/api/credentials/authentication-redirection.any.serviceworker-expected.txt

This is a testharness.js-based test.
PASS getAuthorizationHeaderValue - no redirection
PASS getAuthorizationHeaderValue - same origin redirection
FAIL getAuthorizationHeaderValue - cross origin redirection assert_equals: expected "none" but got "Basic dXNlcjpwYXNz"
Harness: the test ran to completion.
